requirements.txt
hamolicious-vector-class==0.4.10
pygame==2.0.1
